test: split basic materialized view log write tests
